+++ src/main/java/uk/gov/hmcts/reform/prl/models/user/UserRoles.java
@@ -1,5 +1,5 @@
 package uk.gov.hmcts.reform.prl.models.user;
 
 public enum UserRoles {
-    SOLICITOR, CASEWORKER, GATEKEEPER;
+    SOLICITOR, CASEWORKER, GATEKEEPER, JUDGE, COURT_ADMIN, CITIZEN, SYSTEM_UPDATE;
 }
