+++ src/main/java/uk/gov/hmcts/reform/prl/services/LocationRefDataService.java
@@ -6,7 +6,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.prl.clients.LocationRefDataApi;
-import uk.gov.hmcts.reform.prl.enums.YesOrNo;
 import uk.gov.hmcts.reform.prl.models.common.dynamic.DynamicListElement;
 import uk.gov.hmcts.reform.prl.models.court.CourtDetails;
 import uk.gov.hmcts.reform.prl.models.court.CourtVenue;
@@ -77,29 +76,18 @@ public class LocationRefDataService {
         return DynamicListElement.builder().code(key).label(value).build();
     }
 
-    public String getCourtDetailsFromEpimmsId(String baseLocationId, String authToken) {
-        CourtDetails courtDetails = locationRefDataApi.getCourtDetailsByService(authToken,
-                                                                                authTokenGenerator.generate(),
-                                                                                SERVICE_ID);
-        Optional<CourtVenue> courtVenue = courtDetails.getCourtVenues().stream().filter(location -> !"Scotland".equals(location.getRegion()))
+    public Optional<CourtVenue> getCourtDetailsFromEpimmsId(String baseLocationId, String authToken) {
+        CourtDetails courtDetails = locationRefDataApi.getCourtDetailsByService(
+            authToken,
+            authTokenGenerator.generate(),
+            SERVICE_ID
+        );
+        Optional<CourtVenue> courtVenue = courtDetails.getCourtVenues().stream().filter(location -> !"Scotland".equals(
+                location.getRegion()))
             .filter(location -> FAMILY_COURT_TYPE_ID.equalsIgnoreCase(location.getCourtTypeId()))
             .filter(location -> baseLocationId.equalsIgnoreCase(location.getCourtEpimmsId()))
             .findFirst();
-        return courtVenue.map(venue -> venue.getCourtEpimmsId() + "-" + venue.getRegionId()
-            + "-" + venue.getCourtName() + "-" + venue.getPostcode() + "-" + venue.getRegion()
-            + "-" + venue.getSiteName()).orElse("");
-    }
 
-    public YesOrNo cafcassFlag(String regionId) {
-
-        YesOrNo cafcassFlag = YesOrNo.No; //wales
-
-        int intRegionId = Integer.parseInt(regionId);
-
-        if (intRegionId > 0 && intRegionId < 7) {
-            cafcassFlag = YesOrNo.Yes; //english regions
-        }
-
-        return cafcassFlag;
+        return courtVenue;
     }
 }
